refactor: Extract Subnet tests
